+++ packages/web-core/__tests__/entities/character.entity.spec.ts
@@ -25,3 +25,21 @@ test('should get character fields', () => {
   expect(character.displayName).toEqual(displayName);
   expect(character.assets).toEqual(assets);
 });
+
+test('should convert from proto', () => {
+  const id = v4();
+  const resourceName = v4();
+  const displayName = v4();
+
+  const proto = {
+    agentId: id,
+    brainName: resourceName,
+    givenName: displayName,
+  };
+
+  const character = Character.fromProto(proto);
+
+  expect(character.id).toEqual(id);
+  expect(character.resourceName).toEqual(resourceName);
+  expect(character.displayName).toEqual(displayName);
+});