+++ packages/web-core/src/clients/inworld.client.ts
@@ -1,6 +1,6 @@
 import { CapabilitiesConfiguration } from '../../proto/ai/inworld/engine/configuration/configuration.pb';
 import { CancelResponses } from '../../proto/ai/inworld/packets/packets.pb';
-import { GRPC_HOSTNAME, SCENE_PATTERN } from '../common/constants';
+import { GRPC_HOSTNAME } from '../common/constants';
 import {
   Awaitable,
   Capabilities,
@@ -13,6 +13,11 @@ import {
   OnPhomeneFn,
   User,
 } from '../common/data_structures';
+import {
+  SCENE_HAS_INVALID_FORMAT,
+  STOP_DURATION_NATURAL_NUMBER,
+  STOP_TICKS_NATURAL_NUMBER,
+} from '../common/errors';
 import { HistoryItem } from '../components/history';
 import { GrpcAudioPlayback } from '../components/sound/grpc_audio.playback';
 import { GrpcAudioRecorder } from '../components/sound/grpc_audio.recorder';
@@ -23,6 +28,7 @@ import {
 } from '../entities/continuation/session_continuation.entity';
 import { InworldPacket } from '../entities/packets/inworld_packet.entity';
 import { isNaturalNumber } from '../guard/number';
+import { sceneHasValidFormat } from '../guard/scene';
 import { ConnectionService } from '../services/connection.service';
 import { InworldConnectionService } from '../services/inworld_connection.service';
 
@@ -267,21 +273,19 @@ export class InworldClient<
       throw Error('Scene name is required');
     }
 
-    if (!SCENE_PATTERN.test(this.scene)) {
-      throw Error('Scene name has wrong format');
+    if (!sceneHasValidFormat(this.scene)) {
+      throw Error(SCENE_HAS_INVALID_FORMAT);
     }
 
     const { audioPlayback } = this.config;
 
     if (audioPlayback?.stop) {
       if (!isNaturalNumber(audioPlayback.stop.duration)) {
-        throw Error(
-          'Stop duration for audio playback should be a natural number',
-        );
+        throw Error(STOP_DURATION_NATURAL_NUMBER);
       }
 
       if (!isNaturalNumber(audioPlayback.stop.ticks)) {
-        throw Error('Stop ticks for audio playback should be a natural number');
+        throw Error(STOP_TICKS_NATURAL_NUMBER);
       }
     }
   }