+++ packages/web-core/src/common/constants.ts
@@ -4,3 +4,5 @@ export const DEFAULT_USER_NAME = 'User';
 export const DEFAULT_PLAYBACK_SAMPLE_RATE = 16000;
 export const SCENE_PATTERN =
   /^workspaces\/([a-z0-9-_]{1,61})\/(characters|scenes)\/([a-z0-9-_]{0,61})$/iu;
+export const CHARACTER_PATTERN =
+  /^workspaces\/([a-z0-9-_]{1,61})\/characters\/([a-z0-9-_]{0,61})$/iu;