+++ packages/web-core/src/entities/packets/inworld_packet.entity.ts
@@ -1,11 +1,14 @@
 import {
+  Agent,
   DataChunkDataType,
   InworldPacket as ProtoPacket,
+  LoadCharactersCharacterName,
 } from '../../../proto/ai/inworld/packets/packets.pb';
 import {
   InworlControlType,
   InworldPacketType,
 } from '../../common/data_structures';
+import { Character } from '../character.entity';
 import { AudioEvent } from './audio.entity';
 import { CancelResponsesEvent } from './cancel_responses.entity';
 import { ControlEvent } from './control.entity';
@@ -28,8 +31,16 @@ export interface InworldPacketProps {
   silence?: SilenceEvent;
   text?: TextEvent;
   narratedAction?: NarratedAction;
+  sceneMutation?: SceneMutation;
   date: string;
   type: InworldPacketType;
+}
+
+export interface SceneMutation {
+  name?: string;
+  characterNames?: string[];
+  loadedCharacters?: Character[];
+  addedCharacters?: Character[];
 }
 
 export class InworldPacket {
@@ -48,6 +59,7 @@ export class InworldPacket {
   readonly silence: SilenceEvent;
   readonly narratedAction: NarratedAction;
   readonly cancelResponses: CancelResponsesEvent;
+  readonly sceneMutation: SceneMutation;
 
   constructor(props: InworldPacketProps) {
     this.packetId = props.packetId;
@@ -85,6 +97,10 @@ export class InworldPacket {
 
     if (this.isNarratedAction()) {
       this.narratedAction = props.narratedAction;
+    }
+
+    if (this.isSceneMutationResponse() || this.isSceneMutationRequest()) {
+      this.sceneMutation = props.sceneMutation;
     }
   }
 
@@ -159,6 +175,14 @@ export class InworldPacket {
     return this.type === InworldPacketType.NARRATED_ACTION;
   }
 
+  isSceneMutationRequest() {
+    return this.type === InworldPacketType.SCENE_MUTATION_REQUEST;
+  }
+
+  isSceneMutationResponse() {
+    return this.type === InworldPacketType.SCENE_MUTATION_RESPONSE;
+  }
+
   static fromProto(proto: ProtoPacket): InworldPacket {
     const type = this.getType(proto);
 
@@ -191,6 +215,33 @@ export class InworldPacket {
       ...(type === InworldPacketType.NARRATED_ACTION && {
         narratedAction: NarratedAction.fromProto(proto.action),
       }),
+      ...([
+        InworldPacketType.SCENE_MUTATION_REQUEST,
+        InworldPacketType.SCENE_MUTATION_RESPONSE,
+      ].includes(type) && {
+        sceneMutation: {
+          ...(proto.mutation?.loadScene && {
+            name: proto.mutation.loadScene.name,
+          }),
+          ...(proto.mutation?.loadCharacters && {
+            characterNames: proto.mutation.loadCharacters.name.map(
+              (c: LoadCharactersCharacterName) => c.name,
+            ),
+          }),
+          ...(proto.sessionControlResponse?.loadedScene && {
+            loadedCharacters:
+              proto.sessionControlResponse.loadedScene.agents.map(
+                (agent: Agent) => Character.fromProto(agent),
+              ),
+          }),
+          ...(proto.sessionControlResponse?.loadedCharacters && {
+            addedCharacters:
+              proto.sessionControlResponse.loadedCharacters.agents.map(
+                (agent: Agent) => Character.fromProto(agent),
+              ),
+          }),
+        },
+      }),
     });
   }
 
@@ -217,6 +268,13 @@ export class InworldPacket {
       return InworldPacketType.CANCEL_RESPONSE;
     } else if (packet.action?.narratedAction) {
       return InworldPacketType.NARRATED_ACTION;
+    } else if (
+      packet.sessionControlResponse?.loadedScene ||
+      packet.sessionControlResponse?.loadedCharacters
+    ) {
+      return InworldPacketType.SCENE_MUTATION_RESPONSE;
+    } else if (packet.mutation?.loadScene || packet.mutation?.loadCharacters) {
+      return InworldPacketType.SCENE_MUTATION_REQUEST;
     } else {
       return InworldPacketType.UNKNOWN;
     }