+++ packages/web-core/src/entities/scene.entity.ts
@@ -2,54 +2,40 @@ import {
   ActorType,
   Agent,
   InworldPacket as ProtoPacket,
-  SessionControlResponseEvent,
+  LoadedScene,
+  SessionHistoryResponse,
 } from '../../proto/ai/inworld/packets/packets.pb';
 import { Character } from './character.entity';
 
 export interface SceneProps {
-  characters: Character[];
+  name: string;
+  characters?: Character[];
   history?: ProtoPacket[];
 }
 
 export class Scene {
-  characters: Character[] = [];
-  history: ProtoPacket[] = [];
+  name: string;
+  characters: Character[];
+  history: ProtoPacket[];
 
   constructor(props: SceneProps) {
-    this.characters = props.characters;
+    this.name = props.name;
+    this.characters = props.characters ?? [];
     this.history = props.history ?? [];
   }
 
-  static serialize(scene: Scene) {
-    return JSON.stringify(scene);
-  }
-
-  static deserialize(json: string) {
-    try {
-      const { characters } = JSON.parse(json) as SceneProps;
-
-      return new Scene({ characters });
-    } catch (e) {}
-  }
-
   static fromProto({
+    name,
     loadedScene,
     sessionHistory,
-  }: SessionControlResponseEvent) {
-    const characters = (loadedScene?.agents ?? []).map((agent: Agent) => {
-      return new Character({
-        id: agent.agentId,
-        resourceName: agent.brainName,
-        displayName: agent.givenName,
-        assets: {
-          avatarImg: agent.characterAssets.avatarImg,
-          avatarImgOriginal: agent.characterAssets.avatarImgOriginal,
-          rpmModelUri: agent.characterAssets.rpmModelUri,
-          rpmImageUriPortrait: agent.characterAssets.rpmImageUriPortrait,
-          rpmImageUriPosture: agent.characterAssets.rpmImageUriPosture,
-        },
-      });
-    });
+  }: {
+    name: string;
+    loadedScene?: LoadedScene;
+    sessionHistory?: SessionHistoryResponse;
+  }) {
+    const characters = (loadedScene?.agents ?? []).map((agent: Agent) =>
+      Character.fromProto(agent),
+    );
 
     const history =
       sessionHistory?.sessionHistoryItems?.reduce(
@@ -85,6 +71,7 @@ export class Scene {
       ) ?? [];
 
     return new Scene({
+      name,
       characters,
       history,
     });