+++ packages/web-core/src/services/inworld_connection.service.ts
@@ -10,6 +10,11 @@ import {
   TriggerParameter,
   TtsPlaybackAction,
 } from '../common/data_structures';
+import {
+  CHARACTER_HAS_INVALID_FORMAT,
+  SCENE_HAS_INVALID_FORMAT,
+  SCENE_NAME_THE_SAME,
+} from '../common/errors';
 import { GrpcAudioPlayback } from '../components/sound/grpc_audio.playback';
 import { GrpcAudioRecorder } from '../components/sound/grpc_audio.recorder';
 import { GrpcWebRtcLoopbackBiDiSession } from '../components/sound/grpc_web_rtc_loopback_bidi.session';
@@ -17,7 +22,10 @@ import { InworldPlayer } from '../components/sound/inworld_player';
 import { InworldRecorder } from '../components/sound/inworld_recorder';
 import { Character } from '../entities/character.entity';
 import { InworldPacket } from '../entities/packets/inworld_packet.entity';
+import { EventFactory } from '../factories/event';
+import { characterHasValidFormat, sceneHasValidFormat } from '../guard/scene';
 import { ConnectionService } from './connection.service';
+import { FeedbackService } from './feedback.service';
 
 interface InworldConnectionServiceProps<
   InworldPacketT extends InworldPacket = InworldPacket,
@@ -31,6 +39,7 @@ interface InworldConnectionServiceProps<
 export class InworldConnectionService<
   InworldPacketT extends InworldPacket = InworldPacket,
 > {
+  readonly feedback: FeedbackService<InworldPacketT>;
   private connection: ConnectionService;
   private grpcAudioPlayer: GrpcAudioPlayback;
 
@@ -40,6 +49,7 @@ export class InworldConnectionService<
   constructor(props: InworldConnectionServiceProps<InworldPacketT>) {
     this.connection = props.connection;
     this.grpcAudioPlayer = props.grpcAudioPlayer;
+    this.feedback = new FeedbackService(props.connection);
 
     this.player = new InworldPlayer({
       grpcAudioPlayer: this.grpcAudioPlayer,
@@ -91,9 +101,11 @@ export class InworldConnectionService<
   }
 
   async getCurrentCharacter() {
-    await this.connection.getCharacters();
+    return this.connection.getCurrentCharacter();
+  }
 
-    return this.connection.getEventFactory().getCurrentCharacter();
+  setCurrentCharacter(character: Character) {
+    return this.connection.setCurrentCharacter(character);
   }
 
   getHistory() {
@@ -106,10 +118,6 @@ export class InworldConnectionService<
 
   getTranscript() {
     return this.connection.getTranscript();
-  }
-
-  setCurrentCharacter(character: Character) {
-    return this.connection.getEventFactory().setCurrentCharacter(character);
   }
 
   async sendText(text: string, params?: SendPacketParams) {
@@ -225,6 +233,36 @@ export class InworldConnectionService<
         .getEventFactory()
         .narratedAction(text, params?.characters),
     );
+  }
+
+  async reloadScene() {
+    return this.connection.send(() =>
+      EventFactory.loadScene(this.connection.getSceneName()),
+    );
+  }
+
+  async changeScene(name: string) {
+    if (!sceneHasValidFormat(name)) {
+      throw Error(SCENE_HAS_INVALID_FORMAT);
+    }
+
+    if (this.connection.getSceneName() === name) {
+      throw Error(SCENE_NAME_THE_SAME);
+    }
+
+    this.connection.setNextSceneName(name);
+
+    return this.connection.send(() => EventFactory.loadScene(name));
+  }
+
+  async addCharacters(names: string[]) {
+    const invalid = names.find((name) => !characterHasValidFormat(name));
+
+    if (invalid) {
+      throw Error(CHARACTER_HAS_INVALID_FORMAT);
+    }
+
+    return this.connection.send(() => EventFactory.loadCharacters(names));
   }
 
   async sendCustomPacket(getPacket: () => ProtoPacket) {